// code-review-management/types/github.types.ts
const issueState = ["open", "closed"] as const;
const side = ["LEFT", "RIGHT"] as const;
const subjectType = ["line", "file"] as const;
const authorAssociation = [
  "COLLABORATOR",
  "CONTRIBUTOR",
  "FIRST_TIMER",
  "FIRST_TIME_CONTRIBUTOR",
  "MANNEQUIN",
  "MEMBER",
  "NONE",
  "OWNER",
] as const;
const repoVisibility = ["public", "private", "internal"] as const;
const fileDiffStatus = [
  "added",
  "removed",
  "modified",
  "renamed",
  "copied",
  "changed",
  "unchanged",
] as const;
const compareCommitsStatus = [
  "diverged",
  "ahead",
  "behind",
  "identical",
] as const;

export const UserSchema = z.object({
  login: z.string(),
  id: z.number(),
  avatar_url: z.string(),
  permissions: z
    .object({
      pull: z.boolean(),
      push: z.boolean(),
      admin: z.boolean(),
      triage: z.boolean().optional(),
      maintain: z.boolean().optional(),
    })
    .optional(),
});

export const UserIdentitySchema = z.object({
  date: z.string(),
  email: z.string(),
  name: z.string(),
});

export const RepoSchema = z.object({
  id: z.number(),
  name: z.string(),
  full_name: z.string(),
  owner: UserSchema,
  html_url: z.string(),
  description: z.string().nullable(),
  created_at: z.string(),
  updated_at: z.string(),
  pushed_at: z.string(),
  stargazers_count: z.number(),
  watchers_count: z.number(),
  open_issues_count: z.number(),
  has_pull_requests: z.boolean(),
  visibility: z.enum(repoVisibility),
});

export const BranchSchema = z.object({
  label: z.string(),
  ref: z.string(),
  sha: z.string(),
  user: UserSchema,
  repo: RepoSchema,
});

// optional fields are not included in either list or search endpoints
export const PullRequestSchema = z.object({
  url: z.string(),
  id: z.number(),
  html_url: z.string(),
  repository_url: z.string().optional(),
  repository_name: z.string().optional(),
  repository_owner: z.string().optional(),
  diff_url: z.string().optional(),
  number: z.number(),
  state: z.enum(issueState),
  locked: z.boolean(),
  title: z.string(),
  user: UserSchema.nullable(),
  body: z.string().nullable(),
  created_at: z.string(),
  updated_at: z.string(),
  closed_at: z.string().nullable(),
  merged_at: z.string().nullish(),
  merge_commit_sha: z.string().nullish(),
  assignees: z.array(UserSchema).nullable(),
  requested_reviewers: z.array(UserSchema).nullish(),
  head: BranchSchema.optional(),
  base: BranchSchema.optional(),
  author_association: z.enum(authorAssociation),
  draft: z.boolean(),
  assignee: UserSchema.nullable(),
  merged: z.boolean().optional(),
  mergeable: z.boolean().nullable().optional(),
  rebaseable: z.boolean().nullable().optional(),
  mergeable_state: z.string().optional(),
  merged_by: UserSchema.nullable().optional(),
  comments: z.number().optional(),
  review_comments: z.number().optional(),
  commits: z.number().optional(),
  additions: z.number().optional(),
  deletions: z.number().optional(),
  changed_files: z.number().optional(),
  pull_request: z
    .object({
      merged_at: z.string().nullable(),
    })
    .optional(),
});